+++ internal/model/common.go
@@ -52,6 +52,7 @@ func getRecommendationQuery(orgID string) *gorm.DB {
 				"recommendation_sets.current_memory_limit_kib, "+
 				"recommendation_sets.monitoring_start_time, "+
 				"recommendation_sets.monitoring_end_time, "+
+				"recommendation_sets.notification_codes::text AS notification_codes, "+
 				"recommendation_sets.recommendations, "+
 				"recommendation_sets.cpu_variation_short_cost_pct, "+
 				"recommendation_sets.cpu_variation_short_performance_pct, "+